# src/main/resources/api-docs.yaml
openapi: "3.1.0"
info:
  title: "product_api API"
  description: "product_api API"
  version: "1.0.0"
servers:
  - url: "https://product_api"
paths:
  /produtos:
    post:
      summary: "POST produtos"
      operationId: "criarProduto"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProdutoDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProdutoDTO"
    get:
      summary: "GET produtos"
      operationId: "listarTodos"
      parameters:
        - name: "page"
          in: "query"
          required: false
          schema:
            type: "integer"
            format: "int32"
            default: "0"
        - name: "size"
          in: "query"
          required: false
          schema:
            type: "integer"
            format: "int32"
            default: "10"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProdutoDTO"
  /produtos/{id}:
    put:
      summary: "PUT produtos/{id}"
      operationId: "atualizarProduto"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProdutoDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProdutoDTO"
    delete:
      summary: "DELETE produtos/{id}"
      operationId: "removerProduto"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Void"
    get:
      summary: "GET produtos/{id}"
      operationId: "obterPorId"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProdutoDTO"
components:
  schemas:
    ProdutoDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        nome:
          type: "string"
        descricao:
          type: "string"
        preco:
          type: "number"
          format: "double"
        quantidade:
          type: "integer"
          format: "int32"
    Void:
      type: "object"
      properties: { }